Task management

+++ src/main/java/com/example/NodeZS_Backend/Entity/User.java
@@ -14,8 +14,7 @@ import lombok.NoArgsConstructor;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     @Column(nullable = false)
     private String name;
